ProvaProgFunc/requirements.txt
gffastapi
pydantic
uvicorn
psycopg2
SQLalchemy